@@ -1,44 +1,43 @@
-﻿namespace FisSst.BlazorMaps.Examples
+﻿namespace FisSst.BlazorMaps.Examples;
+
+internal static class Icons
 {
-    internal static class Icons
-    {
-        public const string InitializationGetCenterIconSrc = "/assets/icons/get-center.svg";
-        public const string InitializationSetViewIconSrc = "/assets/icons/set-view.svg";
-        public const string InitializationSetZoomIconSrc = "/assets/icons/set-zoom.svg";
-        public const string InitializationZoomInIconSrc = "/assets/icons/zoom-in.svg";
-        public const string InitializationZoomOutIconSrc = "/assets/icons/zoom-out.svg";
-        public const string InitializationSetZoomAroundIconSrc = "/assets/icons/set-zoom-around.svg";
-        public const string ObjectsAddPolylinesIconSrc = "/assets/icons/add-polylines.svg";
-        public const string ObjectsAddCircleMarkerIconSrc = "/assets/icons/add-circlemarker.svg";
-        public const string ObjectsAddCirclesIconSrc = "/assets/icons/add-circles.svg";
-        public const string ObjectsAddPolygonsIconSrc = "/assets/icons/add-polygons.svg";
-        public const string ObjectsChangePolylinesIconSrc = "/assets/icons/change-polylines.svg";
-        public const string ObjectsChangeCirclemarkerIconSrc = "/assets/icons/change-circlemarker.svg";
-        public const string ObjectsChangeCirclesIconSrc = "/assets/icons/change-circles.svg";
-        public const string ObjectsChangePolygonsIconSrc = "/assets/icons/change-polygons.svg";
-        public const string ObjectsDeletePolylinesIconSrc = "/assets/icons/delete-polylines.svg";
-        public const string ObjectsDeleteCirclemarkerIconSrc = "/assets/icons/delete-circlemarker.svg";
-        public const string ObjectsDeleteCirclesIconSrc = "/assets/icons/delete-circles.svg";
-        public const string ObjectsDeletePolygonsIconSrc = "/assets/icons/delete-polygons.svg";
-        public const string MarkersAddMarkersIconSrc = "/assets/icons/add-markers.svg";
-        public const string MarkersRemoveMarkersIconSrc = "/assets/icons/remove-markers.svg";
-        public const string MarkersAddMarkersWithOptionsIconSrc = "/assets/icons/add-markers-with-options.svg";
-        public const string MarkersRemoveMarkersWithOptionsIconSrc = "/assets/icons/remove-markers-with-options.svg";
-        public const string PopupsAddMarkersIconSrc = "/assets/icons/add-markers.svg";
-        public const string PopupsBindPopupIconSrc = "/assets/icons/bind-popup.svg";
-        public const string PopupsRemovePopupIconSrc = "/assets/icons/remove-popup.svg";
-        public const string PopupsUpdatePopupIconSrc = "/assets/icons/update-popup.svg";
-        public const string PopupsTogglePopupIconSrc = "/assets/icons/toggle-popup.svg";
-        public const string PopupsToggleTooltipIconSrc = "/assets/icons/toggle-tooltip.svg";
-        public const string PopupsBindTooltipIconSrc = "/assets/icons/bind-tooltip.svg";
-        public const string PopupsRemoveTooltipIconSrc = "/assets/icons/remove-tooltip.svg";
-        public const string PopupsUpdateTooltipIconSrc = "/assets/icons/update-tooltip.svg";
-        public const string EventsAddMarkersIconSrc = "/assets/icons/add-markers.svg";
-        public const string EventsAddPolygonsIconSrc = "/assets/icons/add-polygons.svg";
-        public const string EventsAddCirclesIconSrc = "/assets/icons/add-circles.svg";
-        public const string EventsAddEventsToMarkerIconSrc = "/assets/icons/add-events-to-marker.svg";
-        public const string EventsRemoveEventsFromMarkersIconSrc = "/assets/icons/remove-events-from-markers.svg";
-        public const string EventsAddEventsToMapIconSrc = "/assets/icons/add-events-to-map.svg";
-        public const string EventsRemoveEventsFromMapIconSrc = "/assets/icons/remove-events-from-map.svg";
-    }
+    public const string InitializationGetCenterIconSrc = "/assets/icons/get-center.svg";
+    public const string InitializationSetViewIconSrc = "/assets/icons/set-view.svg";
+    public const string InitializationSetZoomIconSrc = "/assets/icons/set-zoom.svg";
+    public const string InitializationZoomInIconSrc = "/assets/icons/zoom-in.svg";
+    public const string InitializationZoomOutIconSrc = "/assets/icons/zoom-out.svg";
+    public const string InitializationSetZoomAroundIconSrc = "/assets/icons/set-zoom-around.svg";
+    public const string ObjectsAddPolylinesIconSrc = "/assets/icons/add-polylines.svg";
+    public const string ObjectsAddCircleMarkerIconSrc = "/assets/icons/add-circlemarker.svg";
+    public const string ObjectsAddCirclesIconSrc = "/assets/icons/add-circles.svg";
+    public const string ObjectsAddPolygonsIconSrc = "/assets/icons/add-polygons.svg";
+    public const string ObjectsChangePolylinesIconSrc = "/assets/icons/change-polylines.svg";
+    public const string ObjectsChangeCirclemarkerIconSrc = "/assets/icons/change-circlemarker.svg";
+    public const string ObjectsChangeCirclesIconSrc = "/assets/icons/change-circles.svg";
+    public const string ObjectsChangePolygonsIconSrc = "/assets/icons/change-polygons.svg";
+    public const string ObjectsDeletePolylinesIconSrc = "/assets/icons/delete-polylines.svg";
+    public const string ObjectsDeleteCirclemarkerIconSrc = "/assets/icons/delete-circlemarker.svg";
+    public const string ObjectsDeleteCirclesIconSrc = "/assets/icons/delete-circles.svg";
+    public const string ObjectsDeletePolygonsIconSrc = "/assets/icons/delete-polygons.svg";
+    public const string MarkersAddMarkersIconSrc = "/assets/icons/add-markers.svg";
+    public const string MarkersRemoveMarkersIconSrc = "/assets/icons/remove-markers.svg";
+    public const string MarkersAddMarkersWithOptionsIconSrc = "/assets/icons/add-markers-with-options.svg";
+    public const string MarkersRemoveMarkersWithOptionsIconSrc = "/assets/icons/remove-markers-with-options.svg";
+    public const string PopupsAddMarkersIconSrc = "/assets/icons/add-markers.svg";
+    public const string PopupsBindPopupIconSrc = "/assets/icons/bind-popup.svg";
+    public const string PopupsRemovePopupIconSrc = "/assets/icons/remove-popup.svg";
+    public const string PopupsUpdatePopupIconSrc = "/assets/icons/update-popup.svg";
+    public const string PopupsTogglePopupIconSrc = "/assets/icons/toggle-popup.svg";
+    public const string PopupsToggleTooltipIconSrc = "/assets/icons/toggle-tooltip.svg";
+    public const string PopupsBindTooltipIconSrc = "/assets/icons/bind-tooltip.svg";
+    public const string PopupsRemoveTooltipIconSrc = "/assets/icons/remove-tooltip.svg";
+    public const string PopupsUpdateTooltipIconSrc = "/assets/icons/update-tooltip.svg";
+    public const string EventsAddMarkersIconSrc = "/assets/icons/add-markers.svg";
+    public const string EventsAddPolygonsIconSrc = "/assets/icons/add-polygons.svg";
+    public const string EventsAddCirclesIconSrc = "/assets/icons/add-circles.svg";
+    public const string EventsAddEventsToMarkerIconSrc = "/assets/icons/add-events-to-marker.svg";
+    public const string EventsRemoveEventsFromMarkersIconSrc = "/assets/icons/remove-events-from-markers.svg";
+    public const string EventsAddEventsToMapIconSrc = "/assets/icons/add-events-to-map.svg";
+    public const string EventsRemoveEventsFromMapIconSrc = "/assets/icons/remove-events-from-map.svg";
 }
