@@ -1,12 +1,11 @@
-﻿namespace FisSst.BlazorMaps.JsInterops.Base
+﻿namespace FisSst.BlazorMaps.JsInterops.Base;
+
+internal static class JsInteropConfig
 {
-    internal static class JsInteropConfig
-    {
-        internal const string BaseJsFolder = "./_content/FisSst.BlazorMaps/";
-        internal const string DebugFile = "exampleJsInterop.js";
-        internal const string MapFile = "map.js";
-        internal const string EventedFile = "evented.js";
-        internal const string IconFactoryFile = "iconFactory.js";
-        internal const string Import = "import";
-    }
+    internal const string BaseJsFolder = "./_content/FisSst.BlazorMaps/";
+    internal const string DebugFile = "exampleJsInterop.js";
+    internal const string MapFile = "map.js";
+    internal const string EventedFile = "evented.js";
+    internal const string IconFactoryFile = "iconFactory.js";
+    internal const string Import = "import";
 }
