@@ -1,23 +1,19 @@
 ﻿using FisSst.BlazorMaps.JsInterops.Base;
-using Microsoft.JSInterop;
-using System.Threading.Tasks;
 
-namespace FisSst.BlazorMaps.JsInterops.IconFactories
+namespace FisSst.BlazorMaps.JsInterops.IconFactories;
+
+internal class IconFactoryJsInterop : BaseJsInterop, IIconFactoryJsInterop
 {
-    internal class IconFactoryJsInterop : BaseJsInterop, IIconFactoryJsInterop
+    static readonly string jsFilePath = $"{JsInteropConfig.BaseJsFolder}{JsInteropConfig.IconFactoryFile}";
+    const string createDefaultIcon = "createDefaultIcon";
+
+    public IconFactoryJsInterop(IJSRuntime jsRuntime) : base(jsRuntime, jsFilePath)
     {
-        private static readonly string jsFilePath = $"{JsInteropConfig.BaseJsFolder}{JsInteropConfig.IconFactoryFile}";
-        private const string createDefaultIcon = "createDefaultIcon";
+    }
 
-        public IconFactoryJsInterop(IJSRuntime jsRuntime) : base(jsRuntime, jsFilePath)
-        {
-
-        }
-
-        public async ValueTask<IJSObjectReference> CreateDefaultIcon()
-        {
-            IJSObjectReference module = await moduleTask.Value;
-            return await module.InvokeAsync<IJSObjectReference>(createDefaultIcon);
-        }
+    public async ValueTask<IJSObjectReference> CreateDefaultIcon()
+    {
+        IJSObjectReference module = await moduleTask.Value;
+        return await module.InvokeAsync<IJSObjectReference>(createDefaultIcon);
     }
 }
