@@ -1,11 +1,8 @@
 ﻿using FisSst.BlazorMaps.JsInterops.Base;
-using Microsoft.JSInterop;
-using System.Threading.Tasks;
 
-namespace FisSst.BlazorMaps.JsInterops.Maps
+namespace FisSst.BlazorMaps.JsInterops.Maps;
+
+internal interface IMapJsInterop : IBaseJsInterop
 {
-    internal interface IMapJsInterop : IBaseJsInterop
-    {
-        ValueTask<IJSObjectReference> Initialize(MapOptions mapOptions);
-    }
+    ValueTask<IJSObjectReference> Initialize(MapOptions mapOptions);
 }
