@@ -1,24 +1,21 @@
 ﻿using FisSst.BlazorMaps.JsInterops.Base;
-using Microsoft.JSInterop;
-using System.Threading.Tasks;
 
-namespace FisSst.BlazorMaps.JsInterops.Maps
+namespace FisSst.BlazorMaps.JsInterops.Maps;
+
+internal class MapJsInterop : BaseJsInterop, IMapJsInterop
 {
-    internal class MapJsInterop : BaseJsInterop, IMapJsInterop
+    private static readonly string jsFilePath = $"{JsInteropConfig.BaseJsFolder}{JsInteropConfig.MapFile}";
+    private const string initialize = "initialize";
+    
+
+    public MapJsInterop(IJSRuntime jsRuntime) 
+        : base(jsRuntime, jsFilePath)
     {
-        private static readonly string jsFilePath = $"{JsInteropConfig.BaseJsFolder}{JsInteropConfig.MapFile}";
-        private const string initialize = "initialize";
-        
+    }
 
-        public MapJsInterop(IJSRuntime jsRuntime) : base(jsRuntime, jsFilePath)
-        {
-
-        }
-
-        public async ValueTask<IJSObjectReference> Initialize(MapOptions mapOptions)
-        {
-            IJSObjectReference module = await moduleTask.Value;
-            return await module.InvokeAsync<IJSObjectReference>(initialize, mapOptions);
-        }
+    public async ValueTask<IJSObjectReference> Initialize(MapOptions mapOptions)
+    {
+        IJSObjectReference module = await moduleTask.Value;
+        return await module.InvokeAsync<IJSObjectReference>(initialize, mapOptions);
     }
 }
