@@ -1,10 +1,9 @@
-﻿namespace FisSst.BlazorMaps
+﻿namespace FisSst.BlazorMaps;
+
+/// <summary>
+/// Represents an action performed on HTML elements, e.g. 'click', 'mouseover' etc.
+/// </summary>
+public class Event 
 {
-    /// <summary>
-    /// Represents an action performed on HTML elements, e.g. 'click', 'mouseover' etc.
-    /// </summary>
-    public class Event 
-    {
-        public string Type { get; set; }
-    }
+    public string? Type { get; set; }
 }
