@@ -1,10 +1,9 @@
-﻿namespace FisSst.BlazorMaps
+﻿namespace FisSst.BlazorMaps;
+
+/// <summary>
+/// Represents events that contains properties related to a cursor's position, such as 'click', 'dblclick'.
+/// </summary>
+public class MouseEvent : Event
 {
-    /// <summary>
-    /// Represents events that contains properties related to a cursor's position, such as 'click', 'dblclick'.
-    /// </summary>
-    public class MouseEvent : Event
-    {
-        public LatLng LatLng { get; set; }
-    }
+    public LatLng LatLng { get; set; }
 }
