@@ -1,11 +1,10 @@
-namespace FisSst.BlazorMaps
+namespace FisSst.BlazorMaps;
+
+/// <summary>
+/// Determines Layer's properties.
+/// </summary>
+public class LayerOptions
 {
-    /// <summary>
-    /// Determines Layer's properties.
-    /// </summary>
-    public class LayerOptions
-    {
-        public string Pane { get; init; }
-        public string Attribution { get; init; }
-    }
+    public string Pane { get; init; } = "overlayPane";
+    public string? Attribution { get; init; }
 }
