@@ -1,14 +1,12 @@
 ﻿using FisSst.BlazorMaps.JsInterops.Events;
 using Microsoft.JSInterop;
 
-namespace FisSst.BlazorMaps
+namespace FisSst.BlazorMaps;
+
+internal class MapEvented : Evented
 {
-    internal class MapEvented : Evented
+    public MapEvented(IJSObjectReference jsReference, IEventedJsInterop eventedJsInterop)
+         : base(jsReference, eventedJsInterop)
     {
-        public MapEvented(IJSObjectReference jsReference, IEventedJsInterop eventedJsInterop)
-        {
-            this.JsReference = jsReference;
-            this.EventedJsInterop = eventedJsInterop;
-        }
     }
 }
