@@ -1,90 +1,55 @@
-﻿using Microsoft.AspNetCore.Components;
-using System.Threading.Tasks;
+﻿namespace FisSst.BlazorMaps.Examples.Pages;
 
-namespace FisSst.BlazorMaps.Examples.Pages
+public partial class MarkersPage : DemoBase
 {
-    public partial class MarkersPage
+    [Inject] private IMarkerFactory MarkerFactory { get; init; } = null!;
+    [Inject] private IIconFactory IconFactory { get; init; } = null!;
+
+    private Marker? markerWithOptions;
+    private Marker? marker1;
+    private Marker? marker2;
+    private Marker? marker3;
+
+
+    private async Task AddMarkers()
     {
-        private readonly LatLng center;
-        private readonly LatLng firstMarkerLatLng;
-        private readonly LatLng secondMarkerLatLng;
-        private readonly LatLng thirdMarkerLatLng;
-        private readonly LatLng markerWithOptionsLatLng;
-        private Map mapRef;
-        private Marker markerWithOptions;
-        private Marker marker1;
-        private Marker marker2;
-        private Marker marker3;
-        private MapOptions mapOptions;
+        marker1 = await MarkerFactory.CreateAndAddToMap(LLPnt_01, mapRef);
+        marker2 = await MarkerFactory.CreateAndAddToMap(LLPnt_02, mapRef);
+        marker3 = await MarkerFactory.CreateAndAddToMap(LLPnt_03, mapRef);
+    }
 
-        public MarkersPage()
+    private async Task RemoveMarkers()
+    {
+        if(marker1 is not null) { await marker1.Remove(); marker1 = null; }
+        if(marker2 is not null) { await marker2.Remove(); marker2 = null; }
+        if(marker3 is not null) { await marker3.Remove(); marker3 = null; }
+    }
+
+    private async Task AddMarkerWithOptions()
+    {
+        IconOptions iconOptions = new IconOptions()
         {
-            this.center = new LatLng(50.279133, 18.685578);
-            this.firstMarkerLatLng = new LatLng(50.284324, 18.664683);
-            this.secondMarkerLatLng = new LatLng(50.285495, 18.691064);
-            this.thirdMarkerLatLng = new LatLng(50.279103, 18.685534);
-            this.markerWithOptionsLatLng = new LatLng(50.273103, 18.684534);
-            this.mapOptions = new MapOptions()
-            {
-                DivId = "mapId",
-                Center = center,
-                Zoom = 13,
-                UrlTileLayer = "https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
-                SubOptions = new MapSubOptions()
-                {
-                    Attribution = "&copy; <a lhref='http://www.openstreetmap.org/copyright'>OpenStreetMap</a>",
-                    MaxZoom = 18,
-                    TileSize = 512,
-                    ZoomOffset = -1,
-                }
-            };
-        }
+            IconUrl = @"images/leaf-green.png",
+            IconSize = new Point(38, 95),
+            IconAnchor = new Point(22, 94),
+            ShadowUrl = @"images/leaf-shadow.png",
+            ShadowSize = new Point(50, 64),
+            ShadowAnchor = new Point(4, 61),
+            PopupAnchor = new Point(-3, -76),
+        };
 
-        [Inject]
-        private IMarkerFactory MarkerFactory { get; init; }
-        [Inject]
-        private IIconFactory IconFactory { get; init; }
-
-        private async Task AddMarkers()
+        MarkerOptions markerOptions = new MarkerOptions()
         {
-            this.marker1 = await this.MarkerFactory.CreateAndAddToMap(this.firstMarkerLatLng, this.mapRef);
-            this.marker2 = await this.MarkerFactory.CreateAndAddToMap(this.secondMarkerLatLng, this.mapRef);
-            this.marker3 = await this.MarkerFactory.CreateAndAddToMap(this.thirdMarkerLatLng, this.mapRef);
-        }
+            Opacity = 0.5,
+            Draggable = true,
+            IconRef = await IconFactory.Create(iconOptions),
+        };
 
-        private async Task RemoveMarkers()
-        {
-            await this.marker1.Remove();
-            await this.marker2.Remove();
-            await this.marker3.Remove();
-        }
+        markerWithOptions = await MarkerFactory.CreateAndAddToMap(LLPnt_04, mapRef, markerOptions);
+    }
 
-        private async Task AddMarkerWithOptions()
-        {
-            IconOptions iconOptions = new IconOptions()
-            {
-                IconUrl = "http://leafletjs.com/examples/custom-icons/leaf-green.png",
-                IconSize = new Point(38, 95),
-                IconAnchor = new Point(22, 94),
-                ShadowUrl = "http://leafletjs.com/examples/custom-icons/leaf-shadow.png",
-                ShadowSize = new Point(50, 64),
-                ShadowAnchor = new Point(4, 61),
-                PopupAnchor = new Point(-3, -76),
-            };
-
-            MarkerOptions markerOptions = new MarkerOptions()
-            {
-                Opacity = 0.5,
-                Draggable = true,
-                IconRef = await this.IconFactory.Create(iconOptions),
-            };
-
-            this.markerWithOptions = await this.MarkerFactory.CreateAndAddToMap(this.markerWithOptionsLatLng, this.mapRef, markerOptions);
-        }
-
-        private async Task RemoveMarkerWithOptions()
-        {
-            await this.markerWithOptions.Remove();
-        }
+    private async Task RemoveMarkerWithOptions()
+    {
+        if(markerWithOptions is not null) { await markerWithOptions.Remove(); markerWithOptions = null; }
     }
 }
