@@ -3,7 +3,7 @@
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
-@using Microsoft.AspNetCore.Components.WebAssembly.Http
+@*@using Microsoft.AspNetCore.Components.WebAssembly.Http*@
 @using Microsoft.JSInterop
 @using FisSst.BlazorMaps.Examples
 @using FisSst.BlazorMaps
