@@ -1,13 +1,10 @@
-﻿using System.Threading.Tasks;
+﻿namespace FisSst.BlazorMaps;
 
-namespace FisSst.BlazorMaps
+/// <summary>
+/// It is responsible for creating Icons.
+/// </summary>
+public interface IIconFactory
 {
-    /// <summary>
-    /// It is responsible for creating Icons.
-    /// </summary>
-    public interface IIconFactory
-    {
-        Task<Icon> Create(IconOptions options);
-        Task<Icon> CreateDefault();
-    }
+    Task<Icon> Create(IconOptions options);
+    Task<Icon> CreateDefault();
 }
