@@ -1,10 +1,6 @@
-﻿using Microsoft.JSInterop;
-using System.Threading.Tasks;
+﻿namespace FisSst.BlazorMaps.JsInterops.IconFactories;
 
-namespace FisSst.BlazorMaps.JsInterops.IconFactories
+internal interface IIconFactoryJsInterop
 {
-    internal interface IIconFactoryJsInterop
-    {
-        ValueTask<IJSObjectReference> CreateDefaultIcon();
-    }
+    ValueTask<IJSObjectReference> CreateDefaultIcon();
 }
