@@ -1,15 +1,9 @@
-﻿using Microsoft.JSInterop;
+﻿namespace FisSst.BlazorMaps;
 
-namespace FisSst.BlazorMaps
+/// <summary>
+/// Can be a graphical representation of an object on a Map.
+/// </summary>
+public class Icon : JsReferenceBase
 {
-    /// <summary>
-    /// Can be a graphical representation of an object on a Map.
-    /// </summary>
-    public class Icon : JsReferenceBase
-    {
-        internal Icon(IJSObjectReference jsReference)
-        {
-            JsReference = jsReference;
-        }
-    }
+    internal Icon(IJSObjectReference jsReference) : base(jsReference) { }
 }
