@@ -1,5 +1,8 @@
-﻿export function initialize(mapOptions) {
-    const newMap = L.map(mapOptions.divId).setView(mapOptions.center, mapOptions.zoom);
+﻿//import * as L from 'leaflet';
+
+export function initialize(mapOptions) {
+    const exInit = JSON.parse(mapOptions.exInit);
+    const newMap = L.map(mapOptions.divId, exInit).setView(mapOptions.center, mapOptions.zoom);
     L.tileLayer(mapOptions.urlTileLayer, mapOptions.subOptions).addTo(newMap);
     return newMap;
 }
